Avoid extending BaseInput for custom component
This had lead to issues compiling either with JIT or AOT, depending on
which path for import would be used (umd vs. non-umd). The root cause
seems to be that BaseInput is not exported from the ionic-angular module,
and it's not clear if there's any workaround.

Corresponding quesion on SO:
https://stackoverflow.com/questions/59440199/how-to-import-ionic-angular-baseinput-to-work-with-both-jit-and-aot-compilation

Now we just extend the more generic Ion and implement ControlValueAccessor,
but had to dupe some logic from BaseInput.