Fix schedule copying bug
The id of a copied schedule would be (wrongfully) adopted. This was no
problem if there had been no schedule before, as for POST requests, ids
are ignored anyway. But if the target of copying would have content
already, the resulting PUT request would override and "steal" the
schedule from the other user.